+++ AIPScannerConfig/functions/New-AIPRepository.ps1
@@ -62,7 +62,7 @@
     )
 
     begin {
-        Write-PSFMessage -Level Verbose -String 'Add-AIPRerpository.Message1'
+        Write-PSFMessage -Level Host -String 'Add-AIPRerpository.Message1'
     }
 
     process {
@@ -70,7 +70,7 @@
             Get-AIPScannerRepository
         }
         else {
-            Write-PSFMessage -Level Host -String 'Add-AIPRerpository.Message2'
+            Write-PSFMessage -Level Verbose -String 'Add-AIPRerpository.Message2'
         }
 
         if ($CreateShareOnSharePoint) {
@@ -78,9 +78,9 @@
                 #$Cred = Get-Credential
                 #$Credentials = New-Object Microsoft.SharePoint.Client.SharePointOnlineCredentials($Cred.Username, $Cred.Password)
                 Connect-MsolService
-                $SpLocation = (Get-MsolDomain | Where-Object { $_.isInitial }).Name
-                Write-PSFMessage -Level Verbose -String 'Add-AIPRerpository.Message3' -StringValues "http://$SpLocation/documents/"
-                Add-AIPScannerRepository -Path "http://$SpLocation/documents/"
+                \$spLocation = (Get-MsolDomain | Where-Object { $_.isInitial }).Name
+                Write-PSFMessage -Level Verbose -String 'Add-AIPRerpository.Message3' -StringValues "http://\$spLocation/documents/"
+                Add-AIPScannerRepository -Path "http://\$spLocation/documents/"
             }
             catch {
                 Write-PSFMessage -Level Verbose -String 'Add-AIPRerpository.Message4'