+++ AIPScannerConfig/functions/Start-PrerequisiteCheck.ps1
@@ -79,6 +79,7 @@
 
     process {
         if (-NOT (Assert-ElevatedPermission)) { return }
+        Assert-IEEnhancedSC
         $OriginalPreference = $ProgressPreference
         $ProgressPreference = "SilentlyContinue"
         Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message2'
@@ -88,7 +89,7 @@
                 -or ($ServerVersion -like "Microsoft Windows Server 2016*")`
                 -or ($ServerVersion -like "Microsoft Windows Server 2019*")`
                 -or ($ServerVersion -like "Microsoft Windows Server 2022*")) {
-            Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message3' -StringValues $ServerVersion
+            Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message3'
         }
         else {
             Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message4'
@@ -96,7 +97,7 @@
         }
 
         Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message5'
-        $WindowsFeature = Get-WindowsFeature | Where-Object Name -eq "NFS-Client" > Out-Null
+        $WindowsFeature = Get-WindowsFeature | Where-Object Name -eq "NFS-Client"
 
         if (-NOT ($WindowsFeature.Installed)) {
             Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message6'
@@ -158,8 +159,8 @@
 
             if (Get-LocalUser -Name (Get-PSFConfigValue -Fullname AIPScannerConfig.ScannerAccountName) -ErrorAction SilentlyContinue) {
                 Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message19'
+                New-AIPFileShare
                 if (New-AzureTenantAccountAndApplication) {
-                    New-AIPFileShare
                     New-AIPScannerInstall
                 }
                 else {